pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "betocq"
version = "2.4.0"
description = "Better Together Connectivity Quality test suite."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mobly",
  "mobly-wifi",
  "pyyaml",
]

[project.scripts]
betocq_test_suite = "betocq.betocq_test_suite:main"
local_mobly_runner = "betocq.tools.local_mobly_runner:main"

[project.urls]
Homepage = "https://github.com/android/betocq"
